refactor(repo-icon): split file detection
